requirements_complete.txt
# Flask Web Application Dependencies
Flask==2.3.3
Werkzeug==2.3.7
flask_pymongo
flask_bcrypt
flask_cors
pymongo
dnspython
python-dotenv

# AI Integration
google-generativeai

# Pose Detection / AI Module Dependencies
absl-py==2.3.1
attrs==25.3.0
cffi==1.17.1
contourpy==1.3.2
cycler==0.12.1
flatbuffers==25.2.10
fonttools==4.59.2
jax==0.6.2
jaxlib==0.6.2
kiwisolver==1.4.9
matplotlib==3.10.6
mediapipe==0.10.14
ml_dtypes==0.5.3
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
opt_einsum==3.4.0
packaging==25.0
pillow==11.3.0
protobuf==3.20.3
pycparser==2.22
pygame
pyparsing==3.2.3
python-dateutil==2.9.0.post0
scipy==1.15.3
six==1.17.0
sounddevice==0.5.2
